+++ packages/hono-takibi/src/generators/app/register-component/generate-register-component.ts
