+++ packages/hono-takibi/src/generators/hono/app/index.ts
@@ -4,7 +4,7 @@ import { generateDocs } from '../../openapi/docs/generate-docs'
 import { getHandlerImports } from '../handler/import/get-handler-imports'
 import { getRouteMaps } from './helper/get-route-maps'
 import { generateImportHandlers } from '../handler/import/generate-import-handlers'
-import { generateRegisterComponent } from '../../openapi/components/generate-register-component'
+import { generateRegisterComponent } from '../../app/register-component/generate-register-component'
 import { generateImportRoutes } from './generators/generate-import-routes'
 import { generateApplyOpenapiRoutes } from './generators/generate-apply-openapi-routes'
 import { processImportMap } from './helper/process-import-map'
